requirements.txt
lark-parser==0.8.5
